+++ src/agat/agat_sp_add_introns/script.sh
@@ -1,0 +1,11 @@
+#!/bin/bash
+
+set -eo pipefail
+
+## VIASH START
+## VIASH END
+
+agat_sp_add_introns.pl \
+  -f "$par_gff" \
+  -o "$par_output" \
+  ${par_config:+--config "${par_config}"}